seperate out items from main page ui
