+++ src/constants/index.ts
@@ -1,0 +1,9 @@
+export const TABLE_HEADERS = [
+  "First Name",
+  "Last Name",
+  "City",
+  "Degree",
+  "Specialties",
+  "Years of Experience",
+  "Phone Number",
+];
